Set HTTP 201 for store creation and add controller tests
Added @ResponseStatus(HttpStatus.CREATED) to the StoreController's createStore method to ensure it returns a 201 status upon successful store creation. Introduced unit tests for StoreController to verify both successful store creation and validation errors. Changed StoreRepository ID type from Int to Long for better scalability.

+++ src/main/kotlin/com/example/project24/store/StoreRepository.kt
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface StoreRepository : JpaRepository<Store, Int> {
+interface StoreRepository : JpaRepository<Store, Long> {
 }
